All non-empty lines count towards common whitespace
Previously, any whitespace-only lines were turned into an empty string, following Python's `dedent` library function.

But, I don't think this makes sense in all cases. Instead, if a line were to contain anything (even just whitespace), that makes it significant in some way. We should take them into account when determining the common whitespace.

The only exception to this are the lead line and the last line, which may contain whitespace. For these, we will empty the line and remove the newline any newline char.